# api/docs/swagger.yaml
definitions:
...
  models.ListSpecPrices:
    properties:
      count:
        type: integer
      specialization_prices:
        items:
          $ref: '#/definitions/models.SpecPriceModel'
        type: array
    type: object
...
  models.SpecPriceModel:
    properties:
      created_at:
        type: string
      doctor_id:
        type: string
      id:
        type: integer
      offline_price:
        type: number
      online_price:
        type: number
      specialization_id:
        type: integer
      updated_at:
        type: string
    type: object
  models.SpecPriceReq:
    properties:
      doctor_id:
        type: string
      id:
        type: integer
      offline_price:
        type: number
      online_price:
        type: number
      specialization_id:
        type: integer
    type: object
...
  models.Status:
    properties:
      message:
        type: string
    type: object
...
host: localhost:7070
info:
  contact: {}
  description: api-gateway
  title: Clinic system
  version: "1.0"
paths:
...
  /v1/specprice/{id}:
    get:
      consumes:
      - application/json
      description: Get Specialization price
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SpecPriceModel'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: get specialization price by id
      tags:
      - Specialization price
  /v1/specprice/create:
    post:
      consumes:
      - application/json
      description: Create a new specialization price with the provided details
      parameters:
      - description: Create specialization price
        in: body
        name: SpecPriceInfo
        required: true
        schema:
          $ref: '#/definitions/models.SpecPriceReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SpecPriceModel'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: create specialization price
      tags:
      - Specialization price
  /v1/specprice/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete specialization Price
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Status'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: delete specialization price
      tags:
      - Specialization price
  /v1/specprice/update/{id}:
    put:
      consumes:
      - application/json
      description: Update specialization price
      parameters:
      - description: id
        in: path
        name: id
        type: integer
      - description: Update specialization price
        in: body
        name: UserInfo
        required: true
        schema:
          $ref: '#/definitions/models.SpecPriceReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SpecPriceModel'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: update specialization Price
      tags:
      - Specialization price
  /v1/specprices/{page}/{limit}:
    get:
      consumes:
      - application/json
      description: get all specialization prices
      parameters:
      - description: page
        in: path
        name: page
        type: string
      - description: limit
        in: path
        name: limit
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ListSpecPrices'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: get specialization prices' list
      tags:
      - Specialization price
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"